requirements.txt
# Polymarket Arb Bot - Dependencies

# Core (required)
requests>=2.28.0
python-dotenv>=1.0.0

# Web Server (for dashboard)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Async/WebSocket support (for HFT)
aiohttp>=3.9.0
websockets>=12.0
websocket-client>=1.6.0  # For Polymarket Sports WebSocket

# Blockchain/Trading (install when ready for live trading)
# py-clob-client  # Polymarket's official client
# web3>=6.0.0     # Polygon interaction

# Analysis (for white paper)
pandas>=2.0.0
# matplotlib>=3.7.0

# Note: py-clob-client may need to be installed from git:
# pip install git+https://github.com/Polymarket/py-clob-client.git
